// src/api/bus-stop-api/Dto/busesArrivalInfo.ts
export interface BusesArrivalInfo {
  arrmsg1: string // 첫번째 버스 도착시간
  busId1: string // 확인 필요
  vehicleno1: string
  arrmsg2: string // 두번째 버스 도착시간
  busId2: string
  vehicleno2: string
  busRouteAbrv: string // 버스 노선 약어
  busType1: string // 노선 타입
  busType2: string
}